test: enable Menu-integ-test
